Rethinking jitted org and what classes we need for working with symbolic and numerical basis/orbitals
