+++ atomic/editor.py
@@ -9,7 +9,7 @@ from io import StringIO
 from exa.editor import Editor as ExaEditor
 from atomic.universe import Universe
 from atomic.frame import minimal_frame
-from atomic.basis import CartesianGTFOrder, SphericalGTFOrder, lmap
+from atomic.basis import CartesianGTFOrder, SphericalGTFOrder#, lmap
 
 
 class Editor(ExaEditor):
